Added Rinda/DRb ferret server.
DM::FerretAdapter will now accept a reference to a local index (for
use in a single-process environment), or a reference to a remote
index. The remote index is accessed by passing a name to DM.setup
(formatted like sqlite3's ":memory:"), and it is started with the new
bin/ferret command.